# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LiveForge HTTP API
  version: 1.1.0
  description: |
    Source-aligned contract for the management API, GB28181 API, WebRTC
    WHIP/WHEP signaling, and the default internal cluster relay signaling
    paths. Management responses use `{code,message,data?}`. GB28181 handlers
    return direct JSON objects or arrays. WebRTC signaling returns SDP or
    plain-text errors.

    Management and GB28181 routes accept either an opaque bearer token from
    `api.auth` or an authenticated `lf_session` console cookie when management
    credentials are configured. If no management tokens are configured,
    compatibility mode resolves an anonymous admin principal. Health is always
    public. A missing/invalid credential returns 401; an authenticated role
    without the operation's `x-liveforge-permission` returns 403. Enabled
    per-IP rate limiting can return 429 before a handler runs.

    Internal relay signaling paths are configurable through
    `cluster.rtp.signaling_path` and `cluster.gb28181.signaling_path`; this
    document shows their defaults. Node clients read credentials atomically
    for each request, preferring `api.auth.bearer_token`, then the first named
    admin token. See `docs/recipes/cluster-relay-operations.md` before exposing
    these internal paths.
servers:
  - url: http://127.0.0.1:8090
    description: Default local management listener
security:
  - bearerAuth: []
  - consoleSession: []
...
paths:
...
  /api/v1/gb28181/devices:
    get:
      tags: [gb28181]
      operationId: listGB28181Devices
      summary: List registered devices
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: Devices, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBDeviceSummary'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/channels:
    get:
      tags: [gb28181]
      operationId: listGB28181Channels
      summary: List discovered channels
      x-liveforge-permission: gb28181:read
      parameters:
        - {name: device_id, in: query, required: false, schema: {type: string}}
      responses:
        '200': {description: Channels, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBChannel'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
  /api/v1/gb28181/channels/{channelId}/play:
    parameters:
      - {$ref: '#/components/parameters/ChannelId'}
    post:
      tags: [gb28181]
      operationId: startGB28181LiveView
      summary: Start live view
      x-liveforge-permission: gb28181:control
      responses:
        '200': {description: Session created, content: {application/json: {schema: {$ref: '#/components/schemas/GBSessionStart'}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
        '500': {$ref: '#/components/responses/GBInternalError'}
    delete:
      tags: [gb28181]
      operationId: stopGB28181LiveView
      summary: Stop active live-view sessions for the channel
      x-liveforge-permission: gb28181:control
      responses:
        '200': {description: Sessions stopped, content: {application/json: {schema: {$ref: '#/components/schemas/GBStatus'}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
  /api/v1/gb28181/channels/{channelId}/playback:
    parameters:
      - {$ref: '#/components/parameters/ChannelId'}
    post:
      tags: [gb28181]
      operationId: startGB28181Playback
      summary: Start playback for an RFC3339 range
      x-liveforge-permission: gb28181:control
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/GBPlaybackRequest'}}}
      responses:
        '200': {description: Session created, content: {application/json: {schema: {$ref: '#/components/schemas/GBSessionStart'}}}}
        '400': {$ref: '#/components/responses/GBBadRequest'}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
        '500': {$ref: '#/components/responses/GBInternalError'}
    delete:
      tags: [gb28181]
      operationId: stopGB28181Playback
      summary: Stop active playback sessions for the channel
      x-liveforge-permission: gb28181:control
      responses:
        '200': {description: Sessions stopped, content: {application/json: {schema: {$ref: '#/components/schemas/GBStatus'}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '404': {$ref: '#/components/responses/GBNotFound'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/sessions:
    get:
      tags: [gb28181]
      operationId: listGB28181Sessions
      summary: List active media sessions
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: Sessions, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GBSession'}}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
...
  /api/v1/gb28181/test:
    get:
      tags: [gb28181]
      operationId: runGB28181SelfTest
      summary: Run the local GB28181 RTP/PS protocol lab without a remote platform
      x-liveforge-permission: gb28181:read
      responses:
        '200': {description: Self-test report, content: {application/json: {schema: {$ref: '#/components/schemas/ProtocolTestReport'}}}}
        '401': {$ref: '#/components/responses/GBUnauthorized'}
        '403': {$ref: '#/components/responses/GBForbidden'}
        '429': {$ref: '#/components/responses/GBRateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque management token}
    consoleSession: {type: apiKey, in: cookie, name: lf_session}
    streamBearerAuth: {type: http, scheme: bearer, bearerFormat: publish or subscribe token}
    streamToken: {type: apiKey, in: query, name: token}
  parameters:
...
    ChannelId: {name: channelId, in: path, required: true, schema: {type: string}}
...
  responses:
...
    Unavailable: {description: Required module or runtime manager is disabled, closed, or unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    GBBadRequest: {description: Invalid GB28181 request, content: {application/json: {schema: {$ref: '#/components/schemas/GBError'}}}}
    GBUnauthorized: {description: Management middleware returns an envelope; live/playback publish authorization can return a direct GB error., content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/ApiError'}, {$ref: '#/components/schemas/GBError'}]}}}}
    GBForbidden: {description: Principal lacks the required permission; emitted by management middleware., content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    GBNotFound: {description: Device, channel, or session not found, content: {application/json: {schema: {$ref: '#/components/schemas/GBError'}}}}
    GBRateLimited: {description: Per-IP request limit exceeded; emitted by management middleware., content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    GBInternalError: {description: SIP or media setup failed, content: {application/json: {schema: {$ref: '#/components/schemas/GBError'}}}}
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties: {code: {type: integer, minimum: 400}, message: {type: string}}
...
    ProtocolTestCheck: {type: object, required: [name, passed], properties: {name: {type: string}, passed: {type: boolean}, detail: {type: string}}}
    ProtocolTestReport:
      type: object
      required: [protocol, passed, ran_at, duration_ms, checks]
      properties: {protocol: {type: string, enum: [sip, gb28181]}, passed: {type: boolean}, ran_at: {type: string, format: date-time}, duration_ms: {type: integer, minimum: 0, description: Measured local test execution time in milliseconds.}, checks: {type: array, items: {$ref: '#/components/schemas/ProtocolTestCheck'}}}
...
    GBError: {type: object, required: [error], properties: {error: {type: string}}}
    GBStatus: {type: object, required: [status], properties: {status: {type: string}}}
    GBChannel: {type: object, properties: {channel_id: {type: string}, name: {type: string}, manufacturer: {type: string}, status: {type: string}, ptz_type: {type: integer}, latitude: {type: number}, longitude: {type: number}}}
    GBDeviceSummary: {type: object, properties: {device_id: {type: string}, remote_addr: {type: string}, transport: {type: string}, status: {type: string}, registered_at: {type: string, format: date-time}, last_keepalive: {type: string, format: date-time}, channel_count: {type: integer}}}
...
    GBPlaybackRequest: {type: object, required: [start_time, end_time], properties: {start_time: {type: string, format: date-time}, end_time: {type: string, format: date-time}}}
...
    GBSessionStart: {type: object, required: [stream_key, session_id], properties: {stream_key: {type: string}, session_id: {type: string}}}
    GBSession: {type: object, properties: {id: {type: string}, device_id: {type: string}, channel_id: {type: string}, stream_key: {type: string}, direction: {type: string, enum: [inbound, outbound]}, state: {type: string}, local_port: {type: integer}, transport: {type: string}}}
...